+++ code/hirvi/tests/test_cat.cpp
@@ -103,7 +103,7 @@ TEST(cat_must_be_initialized_appropriately, hirvi_cat)
     cat1_object_struct.local_name = "kissa1";
     cat1_object_struct.brain = rest_brain;
     cat1_object_struct.initial_rotate_vectors = { glm::vec3(0.0f, 1.0f, 1.0f), glm::vec3(0.0f, 0.0f, 1.0f) };
-    cat1_object_struct.initial_rotate_angles = { std::numbers::pi, -0.5f * std::numbers::pi };
+    cat1_object_struct.initial_rotate_angles = { static_cast<float>(std::numbers::pi), -0.5f * static_cast<float>(std::numbers::pi) };
     cat1_object_struct.original_scale_vector = glm::vec3(10.0f, 10.0f, 10.0f);
     cat1_object_struct.cartesian_coordinates = glm::vec3(500.00f, 100.00f, 1000.00f);
